feat(app): surface missing-Accessibility state in menu and Settings
A missing Accessibility permission pauses window management, but
the only signal was the menu-bar warning triangle — the quick menu
had no row explaining what was wrong, and a directly-opened
Settings window looked fully normal while every tiling control was
silently inert.

Add two guided-recovery surfaces, both routing to the same
onboarding grant step (the app's one "why + how" explainer):

- Quick menu: a "Window Management Paused…" row at the top,
  shown only while untrusted, mirroring the Config Issues row.
- Settings: a non-dismissible PermissionPausedBanner in the shared
  chrome (every section + the raw Lua editor). Permission state is
  pushed from AppDelegate into SettingsModel — the Settings layer
  never reads AX state directly.

Split the quick-menu construction into StatusItemController+Menu
to keep both files under the 350-line ceiling.

Docs: user-guide recovery-path prose + menu row. en.json regen
(4 keys). ui-designer-vetted (warning-row shape, banner scope,
non-dismissible).
