+++ Sources/KiwiDesk/StatusItemController.swift
@@ -1,15 +1,19 @@
 import AppKit
 import KiwiDeskCore
 
-/// Owns the menu bar item: the quick menu (#68 §3.10) and the
-/// status icon's state machine (§3.7) — permission warning
-/// beats config error beats the active mode's custom icon
-/// beats the standard glyph.
+/// Owns the menu bar item: the status icon's state machine
+/// (§3.7) — permission warning beats config error beats the
+/// active mode's custom icon beats the standard glyph. The quick
+/// menu (#68 §3.10) is built in `StatusItemController+Menu`.
 @MainActor
 final class StatusItemController: NSObject, NSMenuDelegate {
     var onOpenSettings: () -> Void = {}
     var onOpenDashboard: () -> Void = {}
     var onShowConfigIssues: () -> Void = {}
+    /// Opens the guided permission fix (the onboarding wizard at
+    /// its grant step). Wired unconditionally; the *row* that
+    /// invokes it appears only while Accessibility is missing.
+    var onShowAccessibilityHelp: () -> Void = {}
     var onLoadProfile: (String) -> Void = { _ in }
     /// The saved profiles, the active one, and the unreadable
     /// ones, pulled fresh each menu open. Broken entries stay
@@ -32,10 +36,15 @@ final class StatusItemController: NSObject, NSMenuDelegate {
     private let item: NSStatusItem
     private let menu = NSMenu()
     /// Missing-permission warning overrides everything.
-    private var warning = false
+    /// `private(set)`: the menu builder in `+Menu` reads it to add
+    /// the "Window Management Paused…" row, but only `setWarning`
+    /// may mutate it — every write must go through `render()`.
+    private(set) var warning = false
     /// The last config load left issues (§3.7) — a distinct
-    /// badge so the two causes are never confused.
-    private var configError = false
+    /// badge so the two causes are never confused. `private(set)`
+    /// for the same reason as `warning` (drives the Config Issues
+    /// row; mutated only via `setConfigError`).
+    private(set) var configError = false
     /// Active keybinding mode indicator (SF Symbol or emoji);
     /// nil restores the standard KiwiDesk icon.
     private var modeIcon: String?
@@ -169,153 +178,8 @@ final class StatusItemController: NSObject, NSMenuDelegate {
         }
     }
 
-    // MARK: - Quick menu (§3.10)
-
-    /// Rebuilt on every open so the header, profile list, and
-    /// the conditional Config Issues entry are always current.
-    func menuNeedsUpdate(_ menu: NSMenu) {
-        menu.removeAllItems()
-        let profiles = profilesProvider()
-
-        menu.addItem(headerItem(active: profiles.active))
-        menu.addItem(.separator())
-        menu.addItem(layoutItem())
-        menu.addItem(.separator())
-        menu.addItem(switchProfileItem(profiles))
-        menu.addItem(.separator())
-        if configError {
-            let issues = NSMenuItem(
-                title: L(
-                    "menu.config_issues",
-                    "Config Issues…"
-                ),
-                action: #selector(showConfigIssues),
-                keyEquivalent: ""
-            )
-            issues.target = self
-            // A warning triangle makes the entry unmissable in
-            // the quick menu; safe from the §3.7 status-glyph
-            // distinction here because the row is text-labeled.
-            issues.image = symbol("exclamationmark.triangle.fill")
-            menu.addItem(issues)
-            menu.addItem(.separator())
-        }
-        let settings = NSMenuItem(
-            title: L("menu.settings", "Settings…"),
-            action: #selector(openDashboard),
-            keyEquivalent: ","
-        )
-        settings.target = self
-        settings.image = symbol("gearshape")
-        menu.addItem(settings)
-
-        let accessibility = NSMenuItem(
-            title: L(
-                "menu.accessibility_settings",
-                "Accessibility Settings…"
-            ),
-            action: #selector(openSettings),
-            keyEquivalent: ""
-        )
-        accessibility.target = self
-        accessibility.image = symbol("accessibility")
-        menu.addItem(accessibility)
-
-        menu.addItem(.separator())
-        let support = NSMenuItem(
-            title: L("menu.support", "Support KiwiDesk"),
-            action: #selector(openSupport),
-            keyEquivalent: ""
-        )
-        support.target = self
-        support.image = symbol("heart")
-        menu.addItem(support)
-
-        menu.addItem(.separator())
-        let quit = NSMenuItem(
-            title: L("menu.quit", "Quit KiwiDesk"),
-            action: #selector(NSApplication.terminate(_:)),
-            keyEquivalent: "q"
-        )
-        quit.image = symbol("power")
-        menu.addItem(quit)
-    }
-
-    /// The disabled header row: branding slot + which profile
-    /// is live right now (§3.8/§3.10).
-    private func headerItem(active: String?) -> NSMenuItem {
-        let title =
-            active.map {
-                L("menu.header.active", "KiwiDesk — %1$@", $0)
-            }
-            ?? L(
-                "menu.header.no_profile",
-                "KiwiDesk — no profile"
-            )
-        let header = NSMenuItem(
-            title: title,
-            action: nil,
-            keyEquivalent: ""
-        )
-        header.isEnabled = false
-        header.image =
-            BrandAssets.menuBarIcon
-            ?? symbol("rectangle.3.group")
-        return header
-    }
-
-    /// `load_profile` had no quick path before (§3.10): one
-    /// submenu entry per saved profile, checkmark on the
-    /// active one.
-    private func switchProfileItem(
-        _ profiles: (
-            active: String?, all: [String], broken: Set<String>
-        )
-    ) -> NSMenuItem {
-        let parent = NSMenuItem(
-            title: L("menu.switch_profile", "Switch Profile"),
-            action: nil,
-            keyEquivalent: ""
-        )
-        parent.image = symbol("square.stack.3d.up")
-        let submenu = NSMenu()
-        if profiles.all.isEmpty {
-            let empty = NSMenuItem(
-                title: L(
-                    "menu.no_profiles",
-                    "No saved profiles"
-                ),
-                action: nil,
-                keyEquivalent: ""
-            )
-            empty.isEnabled = false
-            submenu.addItem(empty)
-        }
-        for name in profiles.all {
-            // A broken profile carries no action, so the menu
-            // auto-disables it — greyed but still listed (#246).
-            let broken = profiles.broken.contains(name)
-            let entry = NSMenuItem(
-                title: name,
-                action: broken
-                    ? nil : #selector(loadProfile(_:)),
-                keyEquivalent: ""
-            )
-            entry.target = broken ? nil : self
-            entry.state =
-                name == profiles.active ? .on : .off
-            if broken {
-                entry.toolTip = L(
-                    "menu.profile_broken",
-                    "Can't load — see Config Issues"
-                )
-            }
-            submenu.addItem(entry)
-        }
-        parent.submenu = submenu
-        return parent
-    }
-
+    /// Shared SF Symbol → template-image helper for the menu
+    /// builders (`+Menu`, `+Layout`) and `render`.
     func symbol(_ name: String) -> NSImage? {
         let image = NSImage(
             systemSymbolName: name,
@@ -323,27 +187,5 @@ final class StatusItemController: NSObject, NSMenuDelegate {
         )
         image?.isTemplate = true
         return image
-    }
-
-    // MARK: - Actions
-
-    @objc private func openSettings() {
-        onOpenSettings()
-    }
-
-    @objc private func openDashboard() {
-        onOpenDashboard()
-    }
-
-    @objc private func showConfigIssues() {
-        onShowConfigIssues()
-    }
-
-    @objc private func openSupport() {
-        NSWorkspace.shared.open(SupportLinks.koFi)
-    }
-
-    @objc private func loadProfile(_ sender: NSMenuItem) {
-        onLoadProfile(sender.title)
     }
 }
