+++ Sources/KiwiDeskCore/Resources/Locales/en.json
@@ -331,6 +331,8 @@
   "lua_editor.back_to_visual": "Back to visual editor",
   "lua_editor.editing_directly": "Editing init.lua directly.",
   "lua_editor.foreign_code": "This init.lua contains hand-written Lua, so the visual editor is disabled to avoid overwriting it.",
+  "menu.accessibility_paused": "Window Management Paused…",
+  "menu.accessibility_paused.tooltip": "Click to grant Accessibility permission and resume tiling.",
   "menu.accessibility_settings": "Accessibility Settings…",
   "menu.config_issues": "Config Issues…",
   "menu.header.active": "KiwiDesk — %1$@",
@@ -491,6 +493,8 @@
   "scroll_grid.vertical": "Vertical",
   "scroll_grid.wrap_focus": "Wrap focus",
   "settings.layout.drift": "Live: %1$@ — profile has %2$@ (not saved)",
+  "settings.permission_paused": "Window management is paused — KiwiDesk needs Accessibility permission.",
+  "settings.permission_paused.action": "Enable Accessibility…",
   "shortcuts.add": "Add",
   "shortcuts.add_application": "Add application",
   "shortcuts.add_binding": "Add binding",
